# backend/scripts/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Campus Operations AI API",
                 "description":  "API for Campus Operations Intelligence Platform",
                 "version":  "0.0.1"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  [

                                        ]
                     }
                 ],
    "paths":  {
                  "/api/students/{id}":  {
                                             "get":  {
                                                         "tags":  [
                                                                      "student-controller"
                                                                  ],
                                                         "summary":  "Get student",
                                                         "description":  "Get student by id",
                                                         "operationId":  "get",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "id",
                                                                                "in":  "path",
                                                                                "required":  true,
                                                                                "schema":  {
                                                                                               "type":  "integer",
                                                                                               "format":  "int64"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "OK",
                                                                                       "content":  {
                                                                                                       "*/*":  "@{schema=}"
                                                                                                   }
                                                                                   }
                                                                       },
                                                         "security":  [
                                                                          {
                                                                              "bearerAuth":  [

                                                                                             ]
                                                                          }
                                                                      ]
                                                     },
                                             "put":  {
                                                         "tags":  [
                                                                      "student-controller"
                                                                  ],
                                                         "summary":  "Update student",
                                                         "description":  "Update an existing student",
                                                         "operationId":  "update",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "id",
                                                                                "in":  "path",
                                                                                "required":  true,
                                                                                "schema":  {
                                                                                               "type":  "integer",
                                                                                               "format":  "int64"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "requestBody":  {
                                                                             "content":  {
                                                                                             "application/json":  {
                                                                                                                      "schema":  "@{$ref=#/components/schemas/StudentDto}"
                                                                                                                  }
                                                                                         },
                                                                             "required":  true
                                                                         },
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "OK",
                                                                                       "content":  {
                                                                                                       "*/*":  "@{schema=}"
                                                                                                   }
                                                                                   }
                                                                       },
                                                         "security":  [
                                                                          {
                                                                              "bearerAuth":  [

                                                                                             ]
                                                                          }
                                                                      ]
                                                     },
                                             "delete":  {
                                                            "tags":  [
                                                                         "student-controller"
                                                                     ],
                                                            "summary":  "Delete student",
                                                            "description":  "Delete a student by id",
                                                            "operationId":  "delete",
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "id",
                                                                                   "in":  "path",
                                                                                   "required":  true,
                                                                                   "schema":  {
                                                                                                  "type":  "integer",
                                                                                                  "format":  "int64"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "OK"
                                                                                      }
                                                                          },
                                                            "security":  [
                                                                             {
                                                                                 "bearerAuth":  [

                                                                                                ]
                                                                             }
                                                                         ]
                                                        }
                                         },
...
                  "/api/students":  {
                                        "get":  {
                                                    "tags":  [
                                                                 "student-controller"
                                                             ],
                                                    "summary":  "List students",
                                                    "description":  "Returns paginated list of students",
                                                    "operationId":  "list",
                                                    "parameters":  [
                                                                       {
                                                                           "name":  "page",
                                                                           "in":  "query",
                                                                           "required":  false,
                                                                           "schema":  {
                                                                                          "type":  "integer",
                                                                                          "format":  "int32",
                                                                                          "default":  0
                                                                                      }
                                                                       },
                                                                       {
                                                                           "name":  "size",
                                                                           "in":  "query",
                                                                           "required":  false,
                                                                           "schema":  {
                                                                                          "type":  "integer",
                                                                                          "format":  "int32",
                                                                                          "default":  20
                                                                                      }
                                                                       }
                                                                   ],
                                                    "responses":  {
                                                                      "200":  {
                                                                                  "description":  "OK",
                                                                                  "content":  {
                                                                                                  "*/*":  "@{schema=}"
                                                                                              }
                                                                              }
                                                                  },
                                                    "security":  [
                                                                     {
                                                                         "bearerAuth":  [

                                                                                        ]
                                                                     }
                                                                 ]
                                                },
                                        "post":  {
                                                     "tags":  [
                                                                  "student-controller"
                                                              ],
                                                     "summary":  "Create student",
                                                     "description":  "Create a new student record",
                                                     "operationId":  "create",
                                                     "requestBody":  {
                                                                         "content":  {
                                                                                         "application/json":  {
                                                                                                                  "schema":  "@{$ref=#/components/schemas/StudentDto}"
                                                                                                              }
                                                                                     },
                                                                         "required":  true
                                                                     },
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  "OK",
                                                                                   "content":  {
                                                                                                   "*/*":  "@{schema=}"
                                                                                               }
                                                                               }
                                                                   },
                                                     "security":  [
                                                                      {
                                                                          "bearerAuth":  [

                                                                                         ]
                                                                      }
                                                                  ]
                                                 }
                                    },
...
                  "/api/students/search":  {
                                               "get":  {
                                                           "tags":  [
                                                                        "student-controller"
                                                                    ],
                                                           "summary":  "Search students",
                                                           "description":  "Search students by name or student number",
                                                           "operationId":  "search",
                                                           "parameters":  [
                                                                              {
                                                                                  "name":  "q",
                                                                                  "in":  "query",
                                                                                  "required":  true,
                                                                                  "schema":  {
                                                                                                 "type":  "string"
                                                                                             }
                                                                              },
                                                                              {
                                                                                  "name":  "page",
                                                                                  "in":  "query",
                                                                                  "required":  false,
                                                                                  "schema":  {
                                                                                                 "type":  "integer",
                                                                                                 "format":  "int32",
                                                                                                 "default":  0
                                                                                             }
                                                                              },
                                                                              {
                                                                                  "name":  "size",
                                                                                  "in":  "query",
                                                                                  "required":  false,
                                                                                  "schema":  {
                                                                                                 "type":  "integer",
                                                                                                 "format":  "int32",
                                                                                                 "default":  20
                                                                                             }
                                                                              }
                                                                          ],
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "OK",
                                                                                         "content":  {
                                                                                                         "*/*":  "@{schema=}"
                                                                                                     }
                                                                                     }
                                                                         },
                                                           "security":  [
                                                                            {
                                                                                "bearerAuth":  [

                                                                                               ]
                                                                            }
                                                                        ]
                                                       }
                                           },
...
                  "/api/dashboard/summary":  {
                                                 "get":  {
                                                             "tags":  [
                                                                          "dashboard-controller"
                                                                      ],
                                                             "summary":  "Dashboard summary",
                                                             "description":  "Counts of students, rooms, courses, and schedules",
                                                             "operationId":  "summary",
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK",
                                                                                           "content":  {
                                                                                                           "*/*":  "@{schema=}"
                                                                                                       }
                                                                                       }
                                                                           },
                                                             "security":  [
                                                                              {
                                                                                  "bearerAuth":  [

                                                                                                 ]
                                                                              }
                                                                          ]
                                                         }
                                             },
...
              },
    "components":  {
                       "schemas":  {
                                       "StudentDto":  {
                                                          "type":  "object",
                                                          "properties":  {
                                                                             "id":  {
                                                                                        "type":  "integer",
                                                                                        "format":  "int64"
                                                                                    },
                                                                             "studentNumber":  {
                                                                                                   "type":  "string"
                                                                                               },
                                                                             "fullName":  {
                                                                                              "type":  "string"
                                                                                          },
                                                                             "email":  {
                                                                                           "type":  "string"
                                                                                       },
                                                                             "department":  {
                                                                                                "type":  "string"
                                                                                            },
                                                                             "createdAt":  {
                                                                                               "type":  "string",
                                                                                               "format":  "date-time"
                                                                                           }
                                                                         }
                                                      },
...
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}