# backend/src/main/java/com/amr/campus_operations_ai/student/controller/StudentController.java
package com.amr.campus_operations_ai.student.controller;

import org.springframework.data.domain.Page;
import io.swagger.v3.oas.annotations.Operation;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import com.amr.campus_operations_ai.student.dto.StudentDto;
import com.amr.campus_operations_ai.student.entity.Student;
import com.amr.campus_operations_ai.student.service.StudentService;

import jakarta.validation.Valid;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;

@RestController
@RequestMapping("/api/students")
@SecurityRequirement(name = "bearerAuth")
public class StudentController {

    private final StudentService service;

    public StudentController(StudentService service) {
        this.service = service;
    }

    @GetMapping
    @Operation(summary = "List students", description = "Returns paginated list of students")
    public ResponseEntity<Page<Student>> list(@RequestParam(defaultValue = "0") int page,
                                              @RequestParam(defaultValue = "20") int size) {
        return ResponseEntity.ok(service.list(page, size));
    }

    @GetMapping("/search")
    @Operation(summary = "Search students", description = "Search students by name or student number")
    public ResponseEntity<Page<Student>> search(@RequestParam String q,
                                                @RequestParam(defaultValue = "0") int page,
                                                @RequestParam(defaultValue = "20") int size) {
        return ResponseEntity.ok(service.search(q, page, size));
    }

    @GetMapping("/{id}")
    @Operation(summary = "Get student", description = "Get student by id")
    public ResponseEntity<Student> get(@PathVariable Long id) {
        return service.get(id).map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
    }

    @PostMapping
    @Operation(summary = "Create student", description = "Create a new student record")
    public ResponseEntity<Student> create(@Valid @RequestBody StudentDto dto) {
        return ResponseEntity.ok(service.create(dto));
    }

    @PutMapping("/{id}")
    @Operation(summary = "Update student", description = "Update an existing student")
    public ResponseEntity<Student> update(@PathVariable Long id, @Valid @RequestBody StudentDto dto) {
        return service.update(id, dto).map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
    }

    @DeleteMapping("/{id}")
    @Operation(summary = "Delete student", description = "Delete a student by id")
    public ResponseEntity<Void> delete(@PathVariable Long id) {
        service.delete(id);
        return ResponseEntity.noContent().build();
    }
}